reintroduce PolyLib support
The PolyLib library is available as a submodule for convenience.

The `Makefile.am` and `configure.ac` files have been adapted accordingly
so tat PolyLib support can be easily enabled or disabled using the
`configure` script. Moreover, the `configure` script ensures that
PolyLib and ISL are not enabled at the same time. For now, the default
behaviour remains unchanged and thus ISL is the default library.

New macros 'POLYLIB_SUPPORT' and 'ISL_SUPPORT' have been introduced.

This commit also introduces experimental support for OSL when using the
PolyLib counterpart of CLooG.

Signed-off-by: Harenome Ranaivoarivony Razanajato <razanajato@etu.unistra.fr>